+++ benchmarks/analytic-funnel/server.py
@@ -22,14 +22,14 @@ class Funnel(umbridge.Model):
         s0 = 3
         m1 = 0
         s1 = np.exp(parameters[0][0] / 2)
-        return [[ f(parameters[0][0], m0, s0) + f(parameters[0][1], m1, s1) ]]
+        return [[ float(f(parameters[0][0], m0, s0) + f(parameters[0][1], m1, s1)) ]]
 
     def supports_evaluate(self):
         return True
 
     def gradient(self, out_wrt, in_wrt, parameters, sens, config):
-        return [self.apply_jacobian(out_wrt, in_wrt, parameters, [sens[0], 0], config)[0],
-                self.apply_jacobian(out_wrt, in_wrt, parameters, [0, sens[0]], config)[0]]
+        return [float(self.apply_jacobian(out_wrt, in_wrt, parameters, [sens[0], 0], config)[0]),
+                float(self.apply_jacobian(out_wrt, in_wrt, parameters, [0, sens[0]], config)[0])]
 
     def supports_gradient(self):
         return True
@@ -46,8 +46,8 @@ class Funnel(umbridge.Model):
         m1 = 0
         s1 = np.exp(parameters[0][0] / 2)
 
-        return [vec[1] * dfdx(parameters[0][1], m1, s1)
-              + vec[0] * (dfdx(parameters[0][0], m0, s0) + .5 * s1 * dfds(parameters[0][1], m1, s1))]
+        return [float(vec[1] * dfdx(parameters[0][1], m1, s1)
+              + vec[0] * (dfdx(parameters[0][0], m0, s0) + .5 * s1 * dfds(parameters[0][1], m1, s1)))]
 
     def supports_apply_jacobian(self):
         return True
